split utils into .h and .cpp
